delete resources in each controller
