+++ operator/pkg/controllers/agent/hosted_agent_controller.go
@@ -21,9 +21,11 @@ import (
 	"reflect"
 	"time"
 
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog"
 	"open-cluster-management.io/api/addon/v1alpha1"
+	addonv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -36,6 +38,7 @@ import (
 	globalhubv1alpha4 "github.com/stolostron/multicluster-global-hub/operator/api/operator/v1alpha4"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/config"
 	operatorconstants "github.com/stolostron/multicluster-global-hub/operator/pkg/constants"
+	"github.com/stolostron/multicluster-global-hub/pkg/utils"
 )
 
 type HostedAgentController struct {
@@ -115,17 +118,34 @@ var mghPred = predicate.Funcs{
 
 func (r *HostedAgentController) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	klog.V(2).Infof("Reconcile ClusterManagementAddOn: %v", req.NamespacedName)
-	if !config.GetImportClusterInHosted() {
-		return ctrl.Result{}, nil
-	}
+
 	mgh, err := config.GetMulticlusterGlobalHub(ctx, r.Client)
 	if err != nil {
 		klog.Error("error ", err)
 		return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
 	}
-	if mgh == nil || config.IsPaused(mgh) || mgh.DeletionTimestamp != nil {
+	if mgh == nil || config.IsPaused(mgh) {
 		return ctrl.Result{}, nil
 	}
+	if mgh.DeletionTimestamp != nil || !config.GetImportClusterInHosted() {
+		if config.GetImportClusterInHosted() {
+			hasmanagedHub, err := r.hasManagedHub(ctx)
+			if err != nil {
+				return ctrl.Result{}, err
+			}
+			if hasmanagedHub {
+				klog.Errorf("You need to detach all the managed hub clusters before uninstalling")
+				return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
+			}
+		}
+		err = r.revertClusterManagementAddon(ctx)
+		if err != nil {
+			return ctrl.Result{}, err
+		}
+		err = r.pruneHostedResources(ctx)
+		return ctrl.Result{}, err
+	}
+
 	cma := &v1alpha1.ClusterManagementAddOn{}
 	err = r.Client.Get(ctx, req.NamespacedName, cma)
 	if err != nil {
@@ -143,6 +163,21 @@ func (r *HostedAgentController) Reconcile(ctx context.Context, req ctrl.Request)
 		return ctrl.Result{}, err
 	}
 	return ctrl.Result{}, nil
+}
+
+func (r *HostedAgentController) hasManagedHub(ctx context.Context) (bool, error) {
+	mcaList := &v1alpha1.ManagedClusterAddOnList{}
+	err := r.Client.List(ctx, mcaList)
+	if err != nil {
+		return false, err
+	}
+	for _, mca := range mcaList.Items {
+		if mca.Name == operatorconstants.GHManagedClusterAddonName {
+			return true, nil
+		}
+	}
+
+	return false, nil
 }
 
 // addAddonConfig add the config to cma, will return true if the cma updated
@@ -166,4 +201,59 @@ func addAddonConfig(cma *v1alpha1.ClusterManagementAddOn) bool {
 	cma.Spec.InstallStrategy.Placements = append(cma.Spec.InstallStrategy.Placements,
 		config.GlobalHubHostedAddonPlacementStrategy)
 	return true
+}
+
+func (r *HostedAgentController) revertClusterManagementAddon(ctx context.Context) error {
+	if !config.IsACMResourceReady() {
+		return nil
+	}
+	cmaList := &v1alpha1.ClusterManagementAddOnList{}
+	err := r.Client.List(ctx, cmaList)
+	if err != nil {
+		return err
+	}
+	for _, cma := range cmaList.Items {
+		if !config.HostedAddonList.Has(cma.Name) {
+			continue
+		}
+		err = r.removeGlobalhubConfig(ctx, cma)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (r *HostedAgentController) removeGlobalhubConfig(ctx context.Context, cma v1alpha1.ClusterManagementAddOn) error {
+	if len(cma.Spec.InstallStrategy.Placements) == 0 {
+		return nil
+	}
+	var desiredPlacementStrategy []addonv1alpha1.PlacementStrategy
+	exist := false
+	for _, pl := range cma.Spec.InstallStrategy.Placements {
+		if reflect.DeepEqual(pl.PlacementRef, config.GlobalHubHostedAddonPlacementStrategy.PlacementRef) {
+			exist = true
+			continue
+		}
+		desiredPlacementStrategy = append(desiredPlacementStrategy, pl)
+	}
+	if !exist {
+		return nil
+	}
+	cma.Spec.InstallStrategy.Placements = desiredPlacementStrategy
+	return r.Client.Update(ctx, &cma)
+}
+
+func (r *HostedAgentController) pruneHostedResources(ctx context.Context) error {
+	addonDeployConfig := &addonv1alpha1.AddOnDeploymentConfig{}
+	if err := r.Client.Get(ctx, types.NamespacedName{
+		Namespace: utils.GetDefaultNamespace(),
+		Name:      "global-hub",
+	}, addonDeployConfig); err != nil {
+		if errors.IsNotFound(err) {
+			return nil
+		}
+		return err
+	}
+	return r.Client.Delete(ctx, addonDeployConfig)
 }
