+++ operator/pkg/controllers/hubofhubs/controller.go
@@ -56,7 +56,6 @@ import (
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/config"
 	operatorconstants "github.com/stolostron/multicluster-global-hub/operator/pkg/constants"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/controllers/agent"
-	"github.com/stolostron/multicluster-global-hub/operator/pkg/controllers/hubofhubs/prune"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/controllers/hubofhubs/webhook"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/controllers/transporter/protocol"
 	"github.com/stolostron/multicluster-global-hub/operator/pkg/utils"
@@ -79,7 +78,6 @@ type GlobalHubReconciler struct {
 	log               logr.Logger
 	upgraded          bool
 	operatorConfig    *config.OperatorConfig
-	pruneReconciler   *prune.PruneReconciler
 	webhookReconciler *webhook.WebhookReconciler
 	imageClient       *imagev1client.ImageV1Client
 }
@@ -95,7 +93,6 @@ func NewGlobalHubReconciler(mgr ctrl.Manager, kubeClient kubernetes.Interface,
 		scheme:            mgr.GetScheme(),
 		recorder:          mgr.GetEventRecorderFor(operatorconstants.GlobalHubControllerName),
 		operatorConfig:    operatorConfig,
-		pruneReconciler:   prune.NewPruneReconciler(mgr.GetClient()),
 		webhookReconciler: webhook.NewWebhookReconciler(mgr),
 		imageClient:       imageClient,
 	}
@@ -607,7 +604,6 @@ func (r *GlobalHubReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		return ctrl.Result{}, nil
 	}
 	var reconcileErr error
-	var needRequeue bool
 	// update status condition
 	defer func() {
 		if reconcileErr != nil {
@@ -635,15 +631,6 @@ func (r *GlobalHubReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		return ctrl.Result{}, nil
 	}
 
-	// prune resources if deleting mgh or metrics is disabled
-	needRequeue, err = r.pruneReconciler.Reconcile(ctx, mgh)
-	if err != nil {
-		reconcileErr = fmt.Errorf("failed to prune Global Hub resources %v", err)
-		return ctrl.Result{}, reconcileErr
-	}
-	if needRequeue {
-		return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
-	}
 	if mgh.GetDeletionTimestamp() != nil {
 		return ctrl.Result{}, nil
 	}
